+++ quiz-client/src/components/Header/Header.js
@@ -12,8 +12,8 @@ import React, { useEffect, useState } from "react";
 import "./Header.css";
 import { Link, useLocation } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import { setUser } from "../../store/auth/auth-slice";
-import { setSearchItems } from "../../store/utils/utils-slice";
+import { setUser } from "../../store/features/authSlice";
+import { setSearchItems } from "../../store/features/utilSlice";
 import { GetReturnTime } from "../../utils/Time2String";
 import { toastConfirm } from "../../utils/sweet-alert";
 import { KEYS } from "../../constants/keys.constant";
