+++ quiz-client/src/components/SideMenu/SideMenu.js
