+++ quiz-client/src/components/StartButton/StartButton.css
