+++ quiz-client/src/index.js
@@ -6,7 +6,7 @@ import { ConfigProvider } from "antd";
 import reportWebVitals from "./reportWebVitals";
 import { BrowserRouter } from "react-router-dom";
 import { Provider } from "react-redux";
-import { store } from "./store";
+import { store } from "./store/store";
 
 const primaryTheme = {
   components: {
