+++ quiz-client/src/pages/PageBrowser/PageBrowser.js
@@ -1,12 +1,12 @@
 import React, { useEffect } from "react";
 import { Outlet } from "react-router-dom";
 import Header from "../../components/Header/Header";
-import NavDivider from "../../components/Nav-Divider/NavDivider";
+import NavDivider from "../../components/Divider/NavDivider";
 import "./PageBrowser.css";
 import { withAuthRequired } from "../../hoc/withAuthRequired";
 import { useDispatch } from "react-redux";
-import { setTopicList } from "../../store/topics/topics-slice";
-import { TopicAPI } from "../../api/topic-api";
+import { setTopicList } from "../../store/features/topicSlice";
+import { TopicAPI } from "../../api/topicApi";
 
 function PageBrowser() {
   const dispatch = useDispatch();
